requirements.txt
django==4.1.7
djangorestframework==3.14.0
django-cors-headers==3.14.0
requests==2.28.2
psycopg2==2.9.5

gunicorn==20.1.0
